Update patch set 2
Patch Set 2:

The goal is to not hardcoded the wifi interface in the wpa_supplicant service in init.rc. The reason is: 1) hardcoded is bad and should be avoided. 2) More important reason is, (in the future) for devices with variable interface or even multiple wifi interfaces, we need the ability to change wifi interface of wpa_supplicant at runtime. Of course I understand this is a generic change, that's why I submit it to AOSP -- the change makes wpa_supplicant service to be able to handle more generic cases.

If you want to know a more deeper insight of the change, this is a real problem we encountered on porting android-x86. For x86 platforms, we have so many different wifi drivers with different interface names. Before froyo branch, the services in init.rc cannot accept parameters. As a result, we have to create different names of wpa services to serve different interfaces. This is really a nightmare of us. I'm so glad to see in froyo you add the ability to pass parameters to a service (by ':' notation). The dhcpcd service has been changed to take the advantage (commit 9363b7d5 in system/core). However, the wpa service need to be changed as well to handle variable interface names.

I know you may not need this change. But Please consider to make AOSP to be more generic since some people do need them.

Patch-set: 2