Update patch set 1
Patch Set 1:

On behalf of Vincent Becker (the author):

Setting a device to AUDIO_DEVICE_NONE is used to unset the audio routes, then
depending on the mode the route will fallback to earpiece or speaker (what you
call "device handset").

Removing the test if (device != AUDIO_DEVICE_NONE) will also set the current
device to AUDIO_DEVICE_NONE and avoid the 'device deadlock' just below with the
test "device == prevDevice". I suppose this solution could work also, but then
there is a need to check if setting mDevice to AUDIO_DEVICE_NONE does not have
any unwanted impact elsewhere in the policy.  That test was initially added by
Google when AUDIO_DEVICE_NONE has been introduced in 4.2.1. It could simplify
the code by also removing our patch and this test. Review and tests need to be
performed.

The original use case that brought us to make the patch was to connect an BT
A2DP device, play some music and disconnect the BT headset during playback.
When resuming playback normally on speaker, no audio was heard (as device was
equal to previous device and routing not forced).

Patch-set: 1