Update patch set 1
Patch Set 1:

It is necessary to keep this patch. After intensive testing, it has been shown that when switching from BT A2DP playback to speaker without that patch, audio would be routed to AUDIO_DEVICE_NONE which means ... silence. This patch helps the setOutputDevice method to remember to force routing in case of a previous unrouting (disconnection/switch off of an output device), so that audio is guaranteed to be routed to the current output device.

Patch-set: 1
Reviewer: vincent becker <1018964@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0