Update patch set 2
Patch Set 2:

Actually the comment should be updated: it is mandatory to send AUDIO_DEVICE_NONE to some HALs that need this indication. So when "force" is true we must send AUDIO_DEVICE_NONE to the HAL.
AUDIO_DEVICE_NONE is not an undefined value.

Patch-set: 2
Reviewer: Eric Laurent <1003259@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0