Update patch set 1
Patch Set 1:

Correct, audio policy is part of mediaserver, which will only run in 32-bit initially

So this is a relatively permanent change (or at least until mediaserver can run in 64-bit, which I expect will be a while).

Whereas I expect the libstagefright blacklisting will be more temporary.

Patch-set: 1