Update patch set 1
Patch Set 1:

We are using it because our HAL is written in C++ and we simply want to reuse the C -> C++ wrapper.  We are based on tinyalsa too.  I suppose this interface is still around (and has been since gingerbread at least) because people are using it.

I would see two ways of proceeding: keep this library useful or remove it altogether and face its users with a choice: fork this wrapper in their HAL or switch to another model.

BTW, the Nexus 4's (mako) HAL uses this interface too and also needs to pass a flag when opening an output stream.  In order to do so, openOutputStream() was hacked (http://androidxref.com/4.4.2_r1/xref/hardware/qcom/audio/legacy/alsa_sound/audio_hw_hal.cpp#569).  If the change we are proposing had been done, this hack would not have been needed.

Patch-set: 1