Update patch set 1
Patch Set 1: Code-Review+2

We can accept contributions from AOSP to maintain this HAL as a convenience as you pointed out but the message is that you cannot expect us to keep it up to date and we may decide to remove it from AOSP in the future.
As you may already have forked the policy manager anyways I think you can also fork the legacy HAL and interface.

Patch-set: 1
Label: Code-Review=+2