Update patch set 1
Patch Set 1:

In current mode.
if supplicant crashes due to segmentation fault.
there is no recovery mechanism to restart the supplicant.
the changes are to restart the supplicant from framework if supplicant connection is not active

this issue is very easy to reproduce.
we can simulate the crash using kill command.

Patch-set: 1