Update patch set 1
Patch Set 1:

yes main stream supplicant registered for SIGHUP SIGINT signals
and when called it will send terminate on both the interfaces to framework. but SIGSEGV is not registered its not feasible to register for this in supplicant. As developer even i prefer to debug segmentation fault, but from enduser/customer prospective there should be a recovery mechanism and any how logcat logs will dump the stack trace for debugging.
please try kill -SIGSEGV command to make it crash then all the commands from framework to supplicant will fail and we should manually  
turn off and turn on wlan to restart supplicant.

Patch-set: 1