Update patch set 3
Patch Set 3:

I still believe that the fact that we have to use this workaround hides a problem somewhere else. 
There is no reason why we would have to force a device change just because previous one was not forced.
The logic should be at a higher level than setOutputDevice.
If you are stuck on a KK based product go ahead and merge this.

If this is for a L product then you should move away from the legacy policy as new features (especially for TV) will not be supported in legacy mode.
The new policy manager is in frameworks/av/services/audiopolicy/AudioPolicyManager.cpp

Patch-set: 3